+++ app/src/main/java/com/rohitthebest/manageyourrenters/ui/fragments/trackMoney/monthlyPayment/MonthlyPaymentFragment.kt
@@ -16,7 +16,7 @@ import com.rohitthebest.manageyourrenters.database.model.apiModels.MonthlyPaymen
 import com.rohitthebest.manageyourrenters.database.model.apiModels.MonthlyPaymentCategory
 import com.rohitthebest.manageyourrenters.databinding.FragmentMonthlyPaymentBinding
 import com.rohitthebest.manageyourrenters.others.Constants
-import com.rohitthebest.manageyourrenters.ui.fragments.trackMoney.CustomMenuItems
+import com.rohitthebest.manageyourrenters.ui.fragments.CustomMenuItems
 import com.rohitthebest.manageyourrenters.ui.viewModels.MonthlyPaymentCategoryViewModel
 import com.rohitthebest.manageyourrenters.ui.viewModels.MonthlyPaymentViewModel
 import com.rohitthebest.manageyourrenters.utils.*
